Update service module layout
