+++ src/app/config/service-navigation.config.ts
@@ -1,12 +1,37 @@
-import { ServiceKey } from "./client-services.config";
+export const SERVICES_CONFIG = {
+  documentation: {
+    key: 'documentation',
+    path: '/service/documentation',
+    label: 'Documentación',
+    tab: 'Documentos',
+  },
+  // rates: {
+  //   key: 'rates',
+  //   path: '/service/rates',
+  //   label: 'Tarifas',
+  //   tab: 'Tarifas',
+  // },
+  // billing: {
+  //   key: 'billing',
+  //   path: '/service/billling',
+  //   label: 'Facturación',
+  //   tab: 'Facturación',
+  // },
+  // rips: {
+  //   key: 'rips',
+  //   path: '/service/rips',
+  //   label: 'Rips',
+  //   tab: 'RIPS',
+  // },
+} as const;
 
-// Configuring access to services by client
-export const clientServicesConfig: Record<number, ServiceKey[]> = {
-  8: ['documentation'],            // Axa
-  // 8: ['documentation', 'rates'],            // Axa
-  // 12: ['documentation', 'billing', 'rips'], // BMI
-  // 13: ['documentation', 'rates', 'billing', 'rips'], // Sura TEST
-};
+export type ServiceKey = keyof typeof SERVICES_CONFIG;
+export type ServiceConfig = (typeof SERVICES_CONFIG)[ServiceKey];
 
-// Services available to any other client
-export const defaultServices: ServiceKey[] = ['documentation'];
+// Order in which services appear in the navigation UI
+export const SERVICES_ORDER: readonly ServiceKey[] = [
+  'documentation',
+  // 'rates',
+  // 'billing',
+  // 'rips',
+];
