+++ src/app/shared/layout/service-navigation/service-navigation.component.ts
@@ -2,8 +2,8 @@ import { CommonModule } from '@angular/common';
 import { AfterViewInit, Component, computed, ElementRef, OnDestroy, OnInit, QueryList, signal, ViewChildren } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { Subject, takeUntil } from 'rxjs';
-import { SERVICES_CONFIG, SERVICES_ORDER } from 'src/app/config/client-services.config';
-import { clientServicesConfig, defaultServices } from 'src/app/config/service-navigation.config';
+import { SERVICES_CONFIG, SERVICES_ORDER } from 'src/app/config/service-navigation.config';
+import { clientServicesConfig, defaultServices } from 'src/app/config/client-services.config';
 import { EventManagerService } from 'src/app/services/events-manager/event-manager.service';
 import { NavigationService } from 'src/app/services/navigation/navigation.service';
 import { BreadcrumbComponent } from '../../components/breadcrumb/breadcrumb.component';
